Create Rasp Pi Project
Collabed with Shauniel Nelson and Mohamed Umair

This project created a security alert system that attached a buzzer, sonic sensor, and 2 LEDs (red and yellow)

This code got an input from the sensor, and would detect if there was an object in front of it. It caused a beep and a light to show on, depending on how close the object was to the sensor.